src/third_party/blink/web_tests/platform/linux/virtual/cache-storage-sequence/external/wpt/service-workers/service-worker/interfaces-sw.https-expected.txt

This is a testharness.js-based test.
PASS Interfaces and attributes in ServiceWorkerGlobalScope
FAIL test setup (cache creation) promise_test: Unhandled rejection with value: object "Duplicate identifier WindowOrWorkerGlobalScope"
FAIL Event constructors assert_equals: FetchEvent.isReload should not exist expected (undefined) undefined but got (boolean) false
PASS xhr is not exposed
PASS URL.createObjectURL is not exposed
Harness: the test ran to completion.
